FixABadBug
Fixing a simple bug that creeped in as .float() calls all across the code... look at dataset.py for the fix... rest all is code cleanup.. and some refactoring...

+++ dataset.py
@@ -5,7 +5,7 @@ import numpy as np
 class NumpyCsvDataSet(Dataset):
     def __init__(self, csv_file, device=None):
         self.csv_file = csv_file
-        self.data = torch.as_tensor(np.loadtxt(csv_file, delimiter=',', dtype=float)).to(device)
+        self.data = torch.as_tensor(np.loadtxt(csv_file, delimiter=',', dtype=np.float32)).to(device)
     
     def __len__(self):
         return len(self.data)
